feat(deps): bump electron from 11.0.4 to 11.1.0
Added Electron DLLs like libGLESv2.dll to symbol server. #26965 (Also in 9, 10, 12)
Fixed an issue that a message box in GTK contains no buttons. #26916 (Also in 10, 12)
Fixed an issue where event.reply could sometimes not deliver a reply to an IPC message when cross-site iframes were present. #26926 (Also in 9, 10, 12)
Fixed an occasional crash on Windows related to NativeViewHost::SetParentAccessible. #26951 (Also in 9, 10, 12)